Build a list of parameters for constructing agents

Generate parameters for constructing agents. Given distributions for ages, sexes, medical conditions, SEIR status, and symptomatic conditions for infectious individuals, generate a list of parameters for constructing a set of agents.

generate_agent_params(
  agent_count,
  age_dist,
  p_female,
  seir_dist,
  p_med_cond,
  p_sympt
)

Arguments

agent_count
The number of agents to generate parameters for.

age_dist
A function to sample from a probability distribution or else a vector or list of ages. The vector will be repeated if necessary to produce enough agents.

p_female
A probability of being female or else a vector or list of sexes ("M" or "F") or logical (TRUE for female, FALSE for male). The vector will be repeated if necessary to produce enough agents.

seir_dist
A function to sample from a probability distribution or else a vector or list of SEIR statuses ("S", "E", "I", or "R"). The vector will be repeated if necessary to produce enough agents.

p_med_cond
A probability of having a medical condition or else a vector or list of medical conditions (logical: TRUE or FALSE). The vector will be repeated if necessary to produce enough agents.

p_sympt
A probability of being symptomatic if infectious, or else a vector or list of symptomatic status if infectious (logical: TRUE or FALSE). The vector will be repeated if necessary to produce enough agents.

Value

A named list of parameters for age, sex, seir, med_cond, and sympt.

Examples

# ADD_EXAMPLES_HERE