FUNCTION_TITLE

FUNCTION_DESCRIPTION

set_agent_probs(agents, transmission_params, progression_params)

Arguments

agents
A data frame describing agents, with columns:

age: Age in years (numeric).

age_bkt: The age bracket (ordered factor).

sex: The sex (factor with levels "M" and "F").

med_cond: Has a medical condition (logical).

sympt: Has symptoms of COVID-19, if infectious (logical).

seir: SEIR status (ordered factor with levels "S", "E", "I", and "R").

id: Unique ID number (integer)

ticks: days since reaching current disease stage (relevant to "E" and "I" status) (integer).

target: target ticks for transition to next compartment (relevant to "E" and "I" status only).

transmission_params
A data frame of parameters for disease transmission probability distributions. Columns are:

age_bkt: The age bracket (ordered factor).

sex: The sex (factor with levels "M" and "F").

med_cond: Has a medical condition (logical).

sympt: Has symptoms of COVID-19, if infectious (logical).

mu_shed: Mean of the shedding parameter distribution.

sigma_shed: Scale (standard deviation) of the shedding parameter distribution.

mu_susc: Mean of the susceptibility parameter distribution

sigma_susc: Scale (standard deviation) of the susceptibility parameter distribution.

progression_params
A data frame of parameters for disease progression probability distributions. Columns are:

age_bkt: The age bracket (ordered factor).

sex: The sex (factor with levels "M" and "F").

med_cond: Has a medical condition (logical).

sympt: Has symptoms of COVID-19, if infectious (logical).

shape_ei: Shape parameter for progression from "E" to "I" compartment.

scale_ei: Scale parameter for progression from "E" to "I" compartment.

shape_ir: Shape parameter for progression from "I" to "R" compartment.

scale_ir: Scale parameter for progression from "I" to "R" compartment.

Value

A data.table with transmission probabilities and progression parameters for each agent. Columns are:

age: Age in years (numeric).

age_bkt: The age bracket (ordered factor).

sex: The sex (factor with levels "M" and "F").

med_cond: Has a medical condition (logical).

sympt: Has symptoms of COVID-19, if infectious (logical).

seir: SEIR status (ordered factor with levels "S", "E", "I", and "R").

id: Unique ID number (integer)

ticks: days since reaching current disease stage (relevant to "E" and "I" status) (integer).

target: target ticks for transition to next compartment (relevant to "E" and "I" status only).

x_shed: Shedding parameter.

x_susc: Susceptibility parameter.

shape_ei: Shape parameter for progression from "E" to "I" compartment.

scale_ei: Scale parameter for progression from "E" to "I" compartment.

shape_ir: Shape parameter for progression from "I" to "R" compartment.

scale_ir: Scale parameter for progression from "I" to "R" compartment.

Examples

# ADD_EXAMPLES_HERE